+++ packages/opencode/src/cli/cmd/tui/app.tsx
@@ -54,12 +54,10 @@ import { Provider } from "@/provider/provider"
 import { ArgsProvider, useArgs, type Args } from "./context/args"
 import open from "open"
 import { writeHeapSnapshot } from "v8"
-import { createOpencodeClient, type OpencodeClient } from "@opencode-ai/sdk/v2"
 import { PromptRefProvider, usePromptRef } from "./context/prompt"
 import { TuiConfigProvider, useTuiConfig } from "./context/tui-config"
 import { TuiConfig } from "@/config/tui"
 import { createTuiApi, TuiPluginRuntime, type RouteMap } from "./plugin"
-import type { TuiPluginApi } from "@opencode-ai/plugin/tui"
 import { FormatError, FormatUnknownError } from "@/cli/error"
 
 async function getTerminalBackgroundColor(): Promise<"dark" | "light"> {
@@ -262,30 +260,6 @@ function App(props: { onSnapshot?: () => Promise<string[]> }) {
   const themeState = useTheme()
   const { theme, mode, setMode, locked, lock, unlock } = themeState
   const sync = useSync()
-  const map = new Map<string, OpencodeClient>()
-  const root = "__local__"
-  const scoped = (workspaceID?: string) => {
-    const key = workspaceID ?? root
-    const hit = map.get(key)
-    if (hit) return hit
-
-    const next = createOpencodeClient({
-      baseUrl: sdk.url,
-      fetch: sdk.fetch,
-      directory: sync.data.path.directory || sdk.directory,
-      experimental_workspaceID: workspaceID,
-    })
-    map.set(key, next)
-    return next
-  }
-  const workspace: TuiPluginApi["workspace"] = {
-    current() {
-      return sdk.workspaceID
-    },
-    set(workspaceID) {
-      sdk.setWorkspace(workspaceID)
-    },
-  }
   const exit = useExit()
   const promptRef = usePromptRef()
   const routes: RouteMap = new Map()
@@ -294,9 +268,6 @@ function App(props: { onSnapshot?: () => Promise<string[]> }) {
     routeRev()
     return routes.get(name)?.at(-1)?.render
   }
-  onCleanup(() => {
-    map.clear()
-  })
 
   const api = createTuiApi({
     command,
@@ -307,14 +278,14 @@ function App(props: { onSnapshot?: () => Promise<string[]> }) {
     route,
     routes,
     bump: () => setRouteRev((x) => x + 1),
+    sdk,
     sync,
     theme: themeState,
     toast,
-    client: () => sdk.client,
-    scopedClient: scoped,
-    workspace,
-    event: sdk.event,
     renderer,
+  })
+  onCleanup(() => {
+    api.dispose()
   })
   const [ready, setReady] = createSignal(false)
   TuiPluginRuntime.init(api)
